+++ Program.cs
@@ -96,7 +96,6 @@ builder.Services.AddSingleton<HttpClient, HttpClient>(provider => httpClient)
                     .AddSingleton<IRocketLaunchLiveAPIClientOptions, RocketLaunchLiveAPIClientOptions>(provider => rocketLaunchLiveApiOptions)
                     .AddSingleton<IDiscordClientOptions, DiscordClientOptions>(provider => discordClientOptions)
                     .AddSingleton<IDiscordClient, Dolores.Clients.Discord.DiscordClient>()
-                    .AddSingleton<IUtility, Utility>()
                     .AddSingleton<INasaOptions, NasaOptions>(provider => nasaApiOptions)
                     .AddSingleton<INasaClient, NasaClient>()
                     .AddSingleton(provider => rocketLaunchOptions)
